+++ src/main/java/com/marcoswolf/crm/reparos/infrastructure/repositories/EquipamentoRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface EquipamentoRepository extends JpaRepository<Equipamento, Integer> {
+public interface EquipamentoRepository extends JpaRepository<Equipamento, Long> {
     // Equipamento
     List<Equipamento> findByNumeroSerieContainingIgnoreCase(String numeroSerie);
     // TipoEquipamento
-    List<Equipamento> findByTipoEquipamento_Id(Integer id);
+    List<Equipamento> findByTipoEquipamento_Id(Long id);
 }
